Completed all code for Module 13 Challenge
Major Changes:
Legend formatting addressed in styles.css instead of the JS code. 
Defined a third map layer for Mapbox's dark style, which we then added to the base map layer. 
Replaced the magenta and purple circle colors with shades of orange and red. 
Changed the color of magnitude 5 (and exactly 5) earthquakes from white to orange.
Added a 6th color (darker red) for earthquakes of at least magnitude 6.